requirements.txt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
